requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.2.0
colorama==0.4.6
diffusers==0.29.2
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.15.4
fsspec==2024.6.1
git-filter-repo==2.47.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.14.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.3
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
zipp==3.19.2
